backend: [refac] modularize dictionary endpoint
Also added some comments
