+++ backend/package.json
@@ -10,5 +10,5 @@
     "@azure/functions": "^4.0.0",
     "htmlparser2": "^9.1.0"
   },
-  "main": "src/functions/*.js"
+  "main": "src/functions/*/index.js"
 }
